Add functionality to connect to server from WebGL
This commit adds the Multiplexer Transport to the Network Manager as well as a Simple Web Transport. With these and SSL enabled, we can connect to our server. Note that the URL has changed as we had to do a bunch of DNS mumbo jumbo.

With this commit I was able to get a version of the game up on itch.io! :)
https://alott.itch.io/moonshot

(@Gus - PM me your itch.io username for access)